+++ eras/babbage/impl/testlib/Test/Cardano/Ledger/Babbage/ImpTest.hs
@@ -47,6 +47,7 @@ import qualified Data.Map.Strict as Map
 import Data.Maybe (isNothing)
 import qualified Data.Sequence.Strict as SSeq
 import qualified Data.Set as Set
+import Data.Typeable (Typeable)
 import GHC.Stack (HasCallStack)
 import Lens.Micro
 import Test.Cardano.Ledger.Alonzo.ImpTest
@@ -67,6 +68,7 @@ instance ShelleyEraImp BabbageEra where
   genRegTxCert = shelleyGenRegTxCert
   genUnRegTxCert = shelleyGenUnRegTxCert
   delegStakeTxCert = shelleyDelegStakeTxCert
+  trySubmitTx = trySubmitTopTx
 
 babbageFixupTx ::
   ( HasCallStack
@@ -152,26 +154,27 @@ produceRefScriptsTx scripts = do
     let txOutZero =
           mkBasicTxOut addr mempty & referenceScriptTxOutL .~ SJust script
     pure $ setMinCoinTxOut pp txOutZero
-  let txBody = mkBasicTxBody & outputsTxBodyL .~ SSeq.fromList (NE.toList txOuts)
-  submitTx (mkBasicTx txBody)
+  submitTopTx $
+    mkBasicTx mkBasicTxBody
+      & bodyTxL . outputsTxBodyL .~ SSeq.fromList (NE.toList txOuts)
 
 mkTxWithRefInputs ::
-  (ShelleyEraImp era, BabbageEraTxBody era) =>
+  (ShelleyEraImp era, BabbageEraTxBody era, Typeable l) =>
   TxIn ->
   NonEmpty TxIn ->
-  Tx TopTx era
+  Tx l era
 mkTxWithRefInputs txIn refIns =
-  mkBasicTx $
-    mkBasicTxBody
-      & referenceInputsTxBodyL .~ Set.fromList (NE.toList refIns)
-      & inputsTxBodyL .~ [txIn]
+  mkBasicTx mkBasicTxBody
+    & bodyTxL . referenceInputsTxBodyL .~ Set.fromList (NE.toList refIns)
+    & bodyTxL . inputsTxBodyL .~ [txIn]
 
 submitTxWithRefInputs ::
   (ShelleyEraImp era, BabbageEraTxBody era) =>
   TxIn ->
   NonEmpty TxIn ->
   ImpTestM era (Tx TopTx era)
-submitTxWithRefInputs txIn refIns = submitTx $ mkTxWithRefInputs txIn refIns
+-- TODO make this work with `submitTx`
+submitTxWithRefInputs txIn refIns = submitTopTx $ mkTxWithRefInputs txIn refIns
 
 class
   ( AlonzoEraImp era